+++ src/main/java/com/barlog/loyaltyapi/dto/LevelInfoDto.java
@@ -4,7 +4,7 @@ import lombok.Data;
 @Data @AllArgsConstructor
 public class LevelInfoDto {
     private int level;
-    private long currentXpInLevel;
-    private long xpForNextLevel;
+    private double currentXpInLevel;
+    private double xpForNextLevel;
     private int progressPercentage;
 }
